% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Calculate ICC Values}
\usage{
icc(sigma_alpha, sigma_gamma, sigma_phi, sigma_epsilon)
}
\arguments{
\item{sigma_alpha}{Standard deviation of cluster-level random effect \eqn{\alpha};}

\item{sigma_gamma}{Standard deviation of period-level random effect \eqn{\gamma};}

\item{sigma_phi}{Standard deviation of subject-level random effect \eqn{\phi};}

\item{sigma_epsilon}{Standard deviation of the residual error term.}
}
\value{
A list of values with within-individual ICC for repeated measures (\code{ICC_a}),
within-period ICC (\code{ICC_w}), and between-period ICC (\code{ICC_b}).
}
\description{
Calculate ICC Values
}
\examples{
icc(sigma_alpha = 1.5, sigma_gamma = 0.0, sigma_phi = 3.0, sigma_epsilon = 4.0)
}
