% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stepped_wedge_design.R
\name{make_stepped_wedge_design}
\alias{make_stepped_wedge_design}
\title{Generate a Balanced Stepped-Wedge Design}
\usage{
make_stepped_wedge_design(i, j, intervention_seq)
}
\arguments{
\item{i}{Number of clusters.}

\item{j}{Number of periods.}

\item{intervention_seq}{Number of intervention sequences for the trial.}
}
\value{
A dataset with the following columns:
\itemize{
\item \code{j}, identifying the distinct periods;
\item \code{s}, identifying the distinct intervention sequences;
\item \code{i}, identifying each cluster;
\item \code{x}, denoting whether a certain cluster \code{i} was treated during a certain period \code{j}.
}
}
\description{
Generate a Balanced Stepped-Wedge Design
}
\examples{
# Trial design with 8 clusters, 4 intervention sequences (thus 2 clusters per
# intervention sequence), 5 time periods:
make_stepped_wedge_design(i = 8, j = 5, intervention_seq = 4)
}
