% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simswjm-package.R
\docType{package}
\name{simswjm-package}
\alias{simswjm}
\alias{simswjm-package}
\title{simswjm: Simulate Data from Stepped Wedge Longitudinal Trials}
\description{
Simulate data from stepped wedge cluster randomised trials with continuous, binary, or counts outcomes. Scenarios with non-informative and informative dropout can be accommodated.
}
\author{
\strong{Maintainer}: Alessandro Gasparini \email{alessandro.gasparini@reddooranalytics.se}

Other contributors:
\itemize{
  \item Red Door Analytics AB [copyright holder]
}

}
\keyword{internal}
