% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simswjm-package.R
\docType{package}
\name{simswjm}
\alias{simswjm}
\title{Simulate Data from Stepped Wedge Longitudinal Trials}
\description{
Tools, utilities, and functions that are useful to simulate data
    from stepped wedge longitudinal trials under a variety of study designs
    and assumptions.
}
\author{
Alessandro Gasparini (alessandro.gasparini@reddooranalytics.se)
}
