% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swtrial_inf.R
\name{swtrial_inf}
\alias{swtrial_inf}
\title{Simulate Data from a Stepped-Wedge Trial with Informative Dropout}
\usage{
swtrial_inf(
  repn,
  N,
  i = 8,
  j = 5,
  intervention_seq = 4,
  deltas,
  betas,
  family,
  sigma_epsilon = 0,
  sigma_alpha = 0,
  sigma_gamma = 0,
  sigma_phi = 0,
  lambda = 0.05,
  p = 2,
  nu = 0,
  omega1 = 0,
  omega2 = 0,
  omega3 = 0
)
}
\arguments{
\item{repn}{Index number for the current dataset, useful for simulations;}

\item{N}{Number of subjects per cluster;}

\item{i}{Number of clusters, passed to [make_stepped_wedge_design()].
Defaults to 8;}

\item{j}{Number of time periods, passed to [make_stepped_wedge_design()].
Defaults to 5;}

\item{intervention_seq}{Number of intervention sequences, passed to [make_stepped_wedge_design()].
Defaults to 4}

\item{deltas}{Treatment effect parameter(s) on the longitudinal outcome.
Must be a vector of length `j - 1`, corresponding to a general time on
treatment model, but if all values are the same then that is equivalent
to a constant intervention effect model;}

\item{betas}{Period effect parameters. Must be a vector of length `j`;}

\item{family}{Distribution of the outcome. Can be one of `"gaussian"`,
`"binomial"`, or `"poisson"` for continuous, binary, count outcomes,
respectively. Default link functions are assumed: identity for continuous
outcomes, logit for binary outcomes, and log for count outcomes;}

\item{sigma_epsilon}{Standard deviation of the residual error term. Only
applies when `family = "gaussian"`, and defaults to 0.0;}

\item{sigma_alpha}{Standard deviation of cluster-level random effect \eqn{\alpha}.
Defaults to 0.0;}

\item{sigma_gamma}{Standard deviation of period-level random effect \eqn{\gamma};
Defaults to 0.0;}

\item{sigma_phi}{Standard deviation of subject-level random effect \eqn{\phi}.
Defaults to 0.0;}

\item{lambda}{Scale parameter of the Weibull baseline hazard function.
Defaults to 0.05;}

\item{p}{Shape parameter of the Weibull baseline hazard.
Defaults to 2;}

\item{nu}{Treatment effect in the survival submodel. Defaults to 0.0;}

\item{omega1}{Association parameter linking the random intercept \eqn{\alpha}
to the survival sub-model. Defaults to 0.0;}

\item{omega2}{Association parameter linking the random intercept \eqn{\gamma}
to the survival sub-model. Defaults to 0.0;}

\item{omega3}{Association parameter linking the random intercept \eqn{\phi}
to the survival sub-model. Defaults to 0.0.}
}
\value{
A simulated dataset from a stepped wedge trial with possibly informative
    dropout.
}
\description{
Simulate Data from a Stepped-Wedge Trial with Informative Dropout
}
\examples{
# Simulate a stepped-wedge trial with a continuous outcome:
swtrial_inf(
  repn = 1, N = 3, i = 4, j = 5, intervention_seq = 4,
  deltas = c(0.5, 0.5, 0.5, 0.5),
  betas = c(0, 0.1, 0.2, 0.3, 0.4),
  family = "gaussian",
  sigma_epsilon = 1,
  sigma_alpha = 2,
  sigma_gamma = 0,
  sigma_phi = 3,
  omega1 = 1, omega3 = 1
)
}
