% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simswjm-package.R
\docType{package}
\name{simswjm-package}
\alias{simswjm}
\alias{simswjm-package}
\title{simswjm: Simulate Data from Stepped Wedge Longitudinal Trials}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Alessandro Gasparini \email{alessandro.gasparini@reddooranalytics.se}

Other contributors:
\itemize{
  \item Red Door Analytics AB [copyright holder]
}

}
\keyword{internal}
